Add Crumbling Attuned Shard
This variant of an Attuned Shard is less costly to create in
a warp-plate, but can only be used once. For the same target
warp-plate it stacks up to 16 (and one can create the full
stack at once when attuning).

Each separate teleport event will consume 1 Crumbling Attuned
Shard when teleporting to the associated destination.

Crumbling Attuned Shards are currently prioritized when
selecting the destination (as opposed to default random picking
of a shard).