Crumbling shard useable, fix roundrobin, centralize destroy
The crumbling shard can be used out of a warp plate
entirely, like a scroll, except it incurs the same cost
as teleporting from a plate.

The destroying of used shards is centralized in
PlayerWaystoneManager, as the plate now explicitly
sets the context.warpItem when triggering the tp.

Also fix the warp plate's round robin / tracking of
last used slot to accommodate crumbling shards being
destroyed on teleport (which could skew the index).