Add a new WarpPlateRecipe type for configurable shard recipes
This new type only applies to attunement items that can be created
within the Warp Plate UI, not extra items that are mapped to a
teleport effect.

If the recipe result count field has a value bigger than the item's
maxStackSize(), then the later is used instead.

The recipe syntax is like:

```
{
  "type": "waystones:warp_plate",
  "result": {
    "item": "waystones:attuned_shard",
    "count": 1
  },
  "inner": {
    "item": "minecraft:flint"
  },
  "outer": {
    "item": "waystones:warp_dust"
  }
}
```

Closes gh-410